pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "openai>=1.99.9",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
    "sift-stack-py>=0.8.3",
    "telegram>=0.0.1",
]
